move ReturnSeeker to traverser.py
